requirements.txt
Flask==3.0.3
python-docx==1.1.2
gunicorn==22.0.0